requirements.txt
python-dotenv
pypdf
requests
pydantic
jupyter

openai
gradio
-e .